# Be/identity-service/src/main/java/com/identity/dto/Response/AuthenticationResponse.java
package com.identity.dto.Response;

import lombok.*;
import lombok.experimental.FieldDefaults;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@FieldDefaults(level = AccessLevel.PRIVATE)
public class AuthenticationResponse {
    String token;
    long expiration;
}