# api-docs-identity-service.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/identity","description":"Generated server url"}],"paths":{"/users/registration":{"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/auth/token":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthenticationResponse"}}}}}}},"/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthenticationResponse"}}}}}}},"/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/auth/introspect":{"post":{"tags":["auth-controller"],"operationId":"introspect","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntrospectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseIntrospectResponse"}}}}}}},"/users/{userEmail}":{"get":{"tags":["user-controller"],"operationId":"getAllUser","parameters":[{"name":"userEmail","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"userEmail","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}},"/users/my-info":{"get":{"tags":["user-controller"],"operationId":"getMyInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/users/list-user":{"get":{"tags":["user-controller"],"operationId":"getAllUser_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUserResponse"}}}}}}}},"components":{"schemas":{"UserCreateRequest":{"type":"object","properties":{"username":{"maxLength":2147483647,"minLength":4,"type":"string"},"email":{"maxLength":2147483647,"minLength":8,"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"},"confirmpassword":{"maxLength":2147483647,"minLength":6,"type":"string"}}},"ApiResponseUserResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"result":{"$ref":"#/components/schemas/UserResponse"}}},"PermissionResponse":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"RoleResponse":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"permissions":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/PermissionResponse"}}}},"UserResponse":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/RoleResponse"}},"enable":{"type":"boolean"}}},"AuthenticationRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"ApiResponseAuthenticationResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"result":{"$ref":"#/components/schemas/AuthenticationResponse"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"},"expiration":{"type":"integer","format":"int64"}}},"RefreshRequest":{"type":"object","properties":{"token":{"type":"string"}}},"LogoutRequest":{"type":"object","properties":{"token":{"type":"string"}}},"ApiResponseVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"result":{"type":"object"}}},"IntrospectRequest":{"type":"object","properties":{"token":{"type":"string"}}},"ApiResponseIntrospectResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"result":{"$ref":"#/components/schemas/IntrospectResponse"}}},"IntrospectResponse":{"type":"object","properties":{"valid":{"type":"boolean"}}},"ApiResponseListUserResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"result":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}},"ApiResponseString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"result":{"type":"string"}}}}}}